+++ vm/src/tests/mod.rs
@@ -36,6 +36,9 @@ mod cairo_pie_test;
 #[cfg(feature = "test_utils")]
 mod skip_instruction_test;
 
+#[cfg(feature = "test_utils")]
+mod cairo_test_suite;
+
 //For simple programs that should just succeed and have no special needs.
 //Checks memory holes == 0
 fn run_program_simple(data: &[u8]) {